ursys: split types-urnet.ts into urnet.d.ts and util-urnet.ts
