+++ _ur/common/class-urnet-transaction.ts
@@ -18,7 +18,7 @@
 
 /// TYPE DECLARATIONS /////////////////////////////////////////////////////////
 /// - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
-import type { NP_Address, NP_Hash } from './types-urnet.ts';
+import type { NP_Hash } from '_ur/_types/urnet.d.ts';
 /// - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
 /** a transaction consists functions used to resolve or reject the transaction
  *  for async operations. Additional data can be stored with the transaction
